telegram-bot/requirements.txt
python-telegram-bot
motor
aiohttp
